pyproject.toml
[project]
name = "multimodal-ai"
version = "0.1.0"
description = "企业文档多模态 RAG（本地开源）- 学习型一条龙项目"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["rag", "multimodal", "enterprise-documents", "nlp", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # 基础
    "numpy>=1.26.4",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "rich>=13.7.1",
    "typer>=0.12.3",
    # 文档解析
    "pymupdf>=1.24.9",
    "pillow>=10.4.0",
    # 向量检索
    "faiss-cpu>=1.8.0.post1",
    # Embedding / 模型（本地开源）
    "transformers>=4.45.0",
    "accelerate>=0.33.0",
    "sentence-transformers>=3.0.1",
    # Web Demo
    "streamlit>=1.37.1",
]

[project.optional-dependencies]
ocr = [
    "paddleocr>=2.7.3",
    # paddlepaddle-gpu 需按你的 CUDA 单独安装
]
reranker = [
    "FlagEmbedding>=1.2.10",  # bge-reranker / bge-m3 的另一套封装
]

[project.scripts]
mrag = "mrag.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
# 如果需要从特定源安装，可以在这里配置

